% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{function.portal}
\alias{function.portal}
\title{Taxonomy of species sampled}
\format{
A data frame with 10 number of columns and 54 number of rows
}
\source{
\url{https://lunasare.github.io/spring2023-data-science/data/portal-species-taxonomy.csv}
}
\usage{
function.portal
}
\description{
A data set containing the taxonomy of species sampled for the Portal Project Data.
We followed the taxonomy from the Open Tree of Life .
}
\keyword{datasets}
